Pull up property to parent schema

+++ crime-hardship/src/main/java/uk/gov/justice/laa/crime/hardship/dto/maat_api/SolicitorCosts.java
@@ -13,7 +13,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class SolicitorCosts {
     private BigDecimal solicitorRate;
-    private BigDecimal solicitorHours;
+    private Integer solicitorHours;
     private BigDecimal solicitorVat;
     private BigDecimal solicitorDisb;
     private BigDecimal solicitorEstTotalCost;
